requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
exceptiongroup==1.2.0
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.5.2
packaging==23.2
pluggy==1.4.0
PyJWT==1.7.1
pytest==8.0.1
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
